+++ src/possess/seed.rs
@@ -3,6 +3,7 @@ use crate::{config, AttributeParseError, Item, CONFIG};
 use config::{ArchetypeHandle, ArchetypeKind};
 use rusoto_dynamodb::AttributeValue;
 use serde::{Deserialize, Serialize};
+use sqlx::types::Type as PgType;
 
 #[derive(Serialize, Deserialize, Default, Debug, Clone, PartialEq)]
 pub struct Seed {
@@ -17,10 +18,10 @@ impl Possessable for Seed {
         PossessionKind::Seed(self)
     }
 }
-#[derive(Serialize, Deserialize, Default, Debug, Clone, PartialEq)]
+#[derive(Serialize, Deserialize, Default, Debug, Clone, PartialEq, PgType)]
 pub struct SeedGrower {
     pub id: String,
-    pub generations: u64,
+    pub generations: u32,
 }
 impl SeedGrower {
     pub fn new(id: String, generations: u64) -> Self {
